mmacos: fix codegen dim-arg ordering; restore tCalib (§0 closed)
ROOT CAUSE of the §0 "heap corruption" bug, finally caught via
gdb backtrace inside the running matlab.unittest:

  do_calib_set_var_elt was allocating
    allocate(zern_modes(max(n_zern, 1)))
  BEFORE reading n_zern from prhs(5).  n_zern was uninitialized
  garbage at the allocate, so the array size was junk; the
  subsequent mxCopyPtrToReal8 then read past the source array
  into adjacent heap memory, corrupting the next allocation
  (which surfaced as `malloc(): invalid size (unsorted)` on
  the next mex call).

  Same bug in do_calib_set_target's wf_zern_modes allocation.

The codegen *intended* to read dim-scalar prhs args first, but
the dim-name extractor used the regex `^[A-Za-z_]\w*$` -- only
bare identifiers.  Dim expressions like `max(n_zern, 1)` slipped
through, so `n_zern` was never added to `dim_names` and got
read in declaration order (last), AFTER the array that depends
on it.

Fix: extract every bare identifier from each dim expression via
`re.findall(r'[A-Za-z_]\w*', d)` and exclude Fortran intrinsics
(max, min, size, kind, len).  Now `max(n_zern,1)` correctly
adds `n_zern` to the dim-names set, and the codegen reads it
before allocating zern_modes.  Audited all 89 emitted dispatchers
-- only the two CALIB setters had the bug; rebuild + regen
+ verifier script confirms zero hits.

Defensive companion: +macos/perturb.m now casts use_local to
double before mmacos call (per the
feedback_mmacos_dispatcher_double_only memory note -- correct
hygiene regardless of whether it triggers a bug).

tCalib restored (9 tests, all green) -- the original deferral
in commit 7b68a19 cited §0; that's now closed.

Test results:
  - gdb backtrace on the original failing repro:
    do_calib_set_var_elt_ +0x... → gen_dispatch_ → mexfunction
    (caught the exact crashing function before the malloc
    arena even noticed)
  - tCalib 9/9 pass under ifx build_release
  - codegen-bug verifier script: zero remaining offenders

+++ mmacos/+macos/perturb.m
@@ -36,6 +36,12 @@ if c == 0.0
 end
 del_base = opts.translation / c;
 
-use_local = strcmp(opts.frame, 'local');
+% Cast logical -> double: mmacos dispatcher reads every arg as
+% REAL(8) via mxGetPr+mxCopyPtrToReal8.  Passing a MATLAB `logical`
+% gives mxGetPr the wrong pointer and mxCopyPtrToReal8 reads
+% adjacent heap memory, corrupting the next allocation.  See
+% feedback_mmacos_dispatcher_double_only memory note + macos
+% PLAN.md §0.
+use_local = double(strcmp(opts.frame, 'local'));
 mmacos('perturb_elt', srf, opts.rotation, del_base, use_local);
 end
